@@ -24,19 +24,22 @@
 // SOFTWARE.
 //
 
-#include <catch2/catch.hpp>
+#include <gtest/gtest.h>
+#include <gmock/gmock.h>
 #include <hinder/exception/exception.h>
 #include <cstdint>
 #include <string>
 
-using namespace Catch::Matchers;
+using ::testing::HasSubstr;
+using ::testing::StartsWith;
+using ::testing::EndsWith;
 using namespace hinder;
 
 // ============================================================================
 // Exception creation and data storage
 // ============================================================================
 
-TEST_CASE("Exception creation with .with() chaining", "[exception]") {
+TEST(Exception, ExceptionCreationWithWithChaining) {
     try {
         throw generic_error()
             .message("Something went wrong")
@@ -44,17 +47,17 @@ TEST_CASE("Exception creation with .with() chaining", "[exception]") {
             .with("errno", 2)
             .with("retries", 3U);
     } catch (generic_error const& e) {
-        CHECK(e.contains("message"));
-        CHECK(e.contains("path"));
-        CHECK(e.contains("errno"));
-        CHECK(e.contains("retries"));
-        CHECK(!e.contains("nonexistent"));
+        EXPECT_TRUE(e.contains("message"));
+        EXPECT_TRUE(e.contains("path"));
+        EXPECT_TRUE(e.contains("errno"));
+        EXPECT_TRUE(e.contains("retries"));
+        EXPECT_FALSE(e.contains("nonexistent"));
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("Storing different value types", "[exception]") {
+TEST(Exception, StoringBoolValues) {
     try {
         throw generic_error()
             .with("bool_val", true)
@@ -63,120 +66,166 @@ TEST_CASE("Storing different value types", "[exception]") {
             .with("double_val", 3.14159)
             .with("string_val", "hello");
     } catch (generic_error const& e) {
-        SECTION("bool values") {
-            auto val = e.get_as<bool>("bool_val");
-            REQUIRE(val.has_value());
-            CHECK(*val == true);
-        }
-        SECTION("signed integer values") {
-            auto val = e.get_as<std::int64_t>("int_val");
-            REQUIRE(val.has_value());
-            CHECK(*val == -42);
-        }
-        SECTION("unsigned integer values") {
-            auto val = e.get_as<std::uint64_t>("uint_val");
-            REQUIRE(val.has_value());
-            CHECK(*val == 42U);
-        }
-        SECTION("double values") {
-            auto val = e.get_as<double>("double_val");
-            REQUIRE(val.has_value());
-            CHECK(*val == Catch::Detail::Approx(3.14159));
-        }
-        SECTION("string values") {
-            auto val = e.get_as<std::string>("string_val");
-            REQUIRE(val.has_value());
-            CHECK(*val == "hello");
-        }
+        auto val = e.get_as<bool>("bool_val");
+        ASSERT_TRUE(val.has_value());
+        EXPECT_EQ(*val, true);
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("Flag-style key with no value", "[exception]") {
+TEST(Exception, StoringSignedIntegerValues) {
+    try {
+        throw generic_error()
+            .with("bool_val", true)
+            .with("int_val", -42)
+            .with("uint_val", 42U)
+            .with("double_val", 3.14159)
+            .with("string_val", "hello");
+    } catch (generic_error const& e) {
+        auto val = e.get_as<std::int64_t>("int_val");
+        ASSERT_TRUE(val.has_value());
+        EXPECT_EQ(*val, -42);
+        return;
+    }
+    FAIL() << "Expected exception to be thrown";
+}
+
+TEST(Exception, StoringUnsignedIntegerValues) {
+    try {
+        throw generic_error()
+            .with("bool_val", true)
+            .with("int_val", -42)
+            .with("uint_val", 42U)
+            .with("double_val", 3.14159)
+            .with("string_val", "hello");
+    } catch (generic_error const& e) {
+        auto val = e.get_as<std::uint64_t>("uint_val");
+        ASSERT_TRUE(val.has_value());
+        EXPECT_EQ(*val, 42U);
+        return;
+    }
+    FAIL() << "Expected exception to be thrown";
+}
+
+TEST(Exception, StoringDoubleValues) {
+    try {
+        throw generic_error()
+            .with("bool_val", true)
+            .with("int_val", -42)
+            .with("uint_val", 42U)
+            .with("double_val", 3.14159)
+            .with("string_val", "hello");
+    } catch (generic_error const& e) {
+        auto val = e.get_as<double>("double_val");
+        ASSERT_TRUE(val.has_value());
+        EXPECT_NEAR(*val, 3.14159, 0.00001);
+        return;
+    }
+    FAIL() << "Expected exception to be thrown";
+}
+
+TEST(Exception, StoringStringValues) {
+    try {
+        throw generic_error()
+            .with("bool_val", true)
+            .with("int_val", -42)
+            .with("uint_val", 42U)
+            .with("double_val", 3.14159)
+            .with("string_val", "hello");
+    } catch (generic_error const& e) {
+        auto val = e.get_as<std::string>("string_val");
+        ASSERT_TRUE(val.has_value());
+        EXPECT_EQ(*val, "hello");
+        return;
+    }
+    FAIL() << "Expected exception to be thrown";
+}
+
+TEST(Exception, FlagstyleKeyWithNoValue) {
     try {
         throw generic_error().with("debug_mode");
     } catch (generic_error const& e) {
-        CHECK(e.contains("debug_mode"));
+        EXPECT_TRUE(e.contains("debug_mode"));
         auto val = e.get("debug_mode");
-        REQUIRE(val.has_value());
-        CHECK(std::holds_alternative<std::monostate>(*val));
+        ASSERT_TRUE(val.has_value());
+        EXPECT_TRUE(std::holds_alternative<std::monostate>(*val));
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("Source location is captured", "[exception]") {
+TEST(Exception, SourceLocationIsCaptured) {
     try {
         throw generic_error().message("test");
     } catch (generic_error const& e) {
         auto const& loc = e.location();
-        CHECK(std::string(loc.file_name()).find("exception_tests.cpp") != std::string::npos);
-        CHECK(loc.line() > 0);
+        EXPECT_NE(std::string(loc.file_name()).find("exception_tests.cpp"), std::string::npos);
+        EXPECT_TRUE(loc.line() > 0);
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
 // ============================================================================
 // Data retrieval
 // ============================================================================
 
-TEST_CASE("Getting values that exist", "[exception]") {
+TEST(Exception, GettingValuesThatExist) {
     try {
         throw generic_error()
             .with("count", 42)
             .with("name", "test");
     } catch (generic_error const& e) {
         auto count = e.get("count");
-        REQUIRE(count.has_value());
-        CHECK(std::holds_alternative<std::int64_t>(*count));
+        ASSERT_TRUE(count.has_value());
+        EXPECT_TRUE(std::holds_alternative<std::int64_t>(*count));
 
         auto count_int = e.get_as<int>("count");
-        REQUIRE(count_int.has_value());
-        CHECK(*count_int == 42);
+        ASSERT_TRUE(count_int.has_value());
+        EXPECT_EQ(*count_int, 42);
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("Getting missing keys", "[exception]") {
+TEST(Exception, GettingMissingKeys) {
     try {
         throw generic_error();
     } catch (generic_error const& e) {
-        CHECK(!e.get("missing").has_value());
-        CHECK(!e.get_as<int>("missing").has_value());
-        CHECK(!e.contains("missing"));
+        EXPECT_FALSE(e.get("missing").has_value());
+        EXPECT_FALSE(e.get_as<int>("missing").has_value());
+        EXPECT_FALSE(e.contains("missing"));
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("Iterating over key-value pairs", "[exception]") {
+TEST(Exception, IteratingOverKeyvaluePairs) {
     try {
         throw generic_error()
             .with("a", 1)
             .with("b", 2)
             .with("c", 3);
     } catch (generic_error const& e) {
-        CHECK(e.size() == 3);
+        EXPECT_EQ(e.size(), 3);
         int count = 0;
         for (auto const& [key, value] : e) {
             (void)key;
             (void)value;
             ++count;
         }
-        CHECK(count == 3);
+        EXPECT_EQ(count, 3);
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
 // ============================================================================
 // Formatting
 // ============================================================================
 
-TEST_CASE("to_string() formatting", "[exception]") {
+TEST(Exception, To_stringFormatting) {
     try {
         throw generic_error()
             .message("Failed to open file")
@@ -185,30 +234,30 @@ TEST_CASE("to_string() formatting", "[exception]") {
     } catch (exception const& e) {
         auto result = to_string(e);
 
-        CHECK_THAT(result, Contains("generic_error"));
-        CHECK_THAT(result, Contains("@"));
-        CHECK_THAT(result, Contains("exception_tests.cpp:"));
-        CHECK_THAT(result, Contains("message: Failed to open file"));
-        CHECK_THAT(result, Contains("path: /etc/config.json"));
-        CHECK_THAT(result, Contains("errno: 2"));
+        EXPECT_THAT(result, HasSubstr("generic_error"));
+        EXPECT_THAT(result, HasSubstr("@"));
+        EXPECT_THAT(result, HasSubstr("exception_tests.cpp:"));
+        EXPECT_THAT(result, HasSubstr("message: Failed to open file"));
+        EXPECT_THAT(result, HasSubstr("path: /etc/config.json"));
+        EXPECT_THAT(result, HasSubstr("errno: 2"));
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("to_string() empty exception", "[exception]") {
+TEST(Exception, To_stringEmptyException) {
     try {
         throw generic_error();
     } catch (exception const& e) {
         auto result = to_string(e);
-        CHECK_THAT(result, Contains("generic_error"));
-        CHECK_THAT(result, Contains("@"));
+        EXPECT_THAT(result, HasSubstr("generic_error"));
+        EXPECT_THAT(result, HasSubstr("@"));
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("to_json() formatting", "[exception]") {
+TEST(Exception, To_jsonFormatting) {
     try {
         throw generic_error()
             .message("Test message")
@@ -217,22 +266,22 @@ TEST_CASE("to_json() formatting", "[exception]") {
     } catch (exception const& e) {
         auto result = to_json(e);
 
-        CHECK_THAT(result, StartsWith("{"));
-        CHECK_THAT(result, EndsWith("}"));
-        CHECK_THAT(result, Contains("\"type\":\"generic_error\""));
-        CHECK_THAT(result, Contains("\"source\":{"));
-        CHECK_THAT(result, Contains("\"file\":"));
-        CHECK_THAT(result, Contains("\"line\":"));
-        CHECK_THAT(result, Contains("\"data\":{"));
-        CHECK_THAT(result, Contains("\"message\":\"Test message\""));
-        CHECK_THAT(result, Contains("\"count\":42"));
-        CHECK_THAT(result, Contains("\"enabled\":true"));
+        EXPECT_THAT(result, StartsWith("{"));
+        EXPECT_THAT(result, EndsWith("}"));
+        EXPECT_THAT(result, HasSubstr("\"type\":\"generic_error\""));
+        EXPECT_THAT(result, HasSubstr("\"source\":{"));
+        EXPECT_THAT(result, HasSubstr("\"file\":"));
+        EXPECT_THAT(result, HasSubstr("\"line\":"));
+        EXPECT_THAT(result, HasSubstr("\"data\":{"));
+        EXPECT_THAT(result, HasSubstr("\"message\":\"Test message\""));
+        EXPECT_THAT(result, HasSubstr("\"count\":42"));
+        EXPECT_THAT(result, HasSubstr("\"enabled\":true"));
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("JSON escaping special characters", "[exception]") {
+TEST(Exception, JsonEscapingSpecialCharacters) {
     try {
         throw generic_error()
             .with("quotes", "say \"hello\"")
@@ -242,89 +291,88 @@ TEST_CASE("JSON escaping special characters", "[exception]") {
     } catch (exception const& e) {
         auto result = to_json(e);
 
-        CHECK_THAT(result, Contains("say \\\"hello\\\""));
-        CHECK_THAT(result, Contains("path\\\\to\\\\file"));
-        CHECK_THAT(result, Contains("line1\\nline2"));
-        CHECK_THAT(result, Contains("col1\\tcol2"));
+        EXPECT_THAT(result, HasSubstr("say \\\"hello\\\""));
+        EXPECT_THAT(result, HasSubstr("path\\\\to\\\\file"));
+        EXPECT_THAT(result, HasSubstr("line1\\nline2"));
+        EXPECT_THAT(result, HasSubstr("col1\\tcol2"));
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("JSON escaping control characters", "[exception]") {
+TEST(Exception, JsonEscapingControlCharacters) {
     try {
         throw generic_error().with("bell", std::string("ring\x07"));
     } catch (exception const& e) {
         auto result = to_json(e);
-        CHECK_THAT(result, Contains("ring\\u0007"));
+        EXPECT_THAT(result, HasSubstr("ring\\u0007"));
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
 // ============================================================================
 // Macros
 // ============================================================================
 
-TEST_CASE("HINDER_THROW captures source location", "[exception]") {
+TEST(Exception, Hinder_throwCapturesSourceLocation) {
     try {
         HINDER_THROW(generic_error).message("test");
     } catch (exception const& e) {
-        CHECK(std::string(e.location().file_name()).find("exception_tests.cpp")
-              != std::string::npos);
+        EXPECT_NE(std::string(e.location().file_name()).find("exception_tests.cpp"), std::string::npos);
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("HINDER_THROW type hierarchy", "[exception]") {
-    CHECK_THROWS_AS(HINDER_THROW(generic_error), generic_error);
-    CHECK_THROWS_AS(HINDER_THROW(generic_error), exception);
-    CHECK_THROWS_AS(HINDER_THROW(generic_error), std::runtime_error);
-    CHECK_THROWS_AS(HINDER_THROW(generic_error), std::exception);
+TEST(Exception, Hinder_throwTypeHierarchy) {
+    EXPECT_THROW(HINDER_THROW(generic_error), generic_error);
+    EXPECT_THROW(HINDER_THROW(generic_error), exception);
+    EXPECT_THROW(HINDER_THROW(generic_error), std::runtime_error);
+    EXPECT_THROW(HINDER_THROW(generic_error), std::exception);
 }
 
-TEST_CASE("HINDER_EXPECTS throws on false", "[exception]") {
+TEST(Exception, Hinder_expectsThrowsOnFalse) {
     try {
         int x = 0;
         HINDER_EXPECTS(x > 0, generic_error);
-        FAIL("Expected exception");
+        FAIL() << "Expected exception";
     } catch (exception const& e) {
-        CHECK(e.get_as<std::string>("condition").value_or("") == "x > 0");
-        CHECK(e.get_as<std::string>("check_type").value_or("") == "precondition");
+        EXPECT_EQ(e.get_as<std::string>("condition").value_or(""), "x > 0");
+        EXPECT_EQ(e.get_as<std::string>("check_type").value_or(""), "precondition");
     }
 }
 
-TEST_CASE("HINDER_EXPECTS passes on true", "[exception]") {
-    CHECK_NOTHROW(HINDER_EXPECTS(true, generic_error));
+TEST(Exception, Hinder_expectsPassesOnTrue) {
+    EXPECT_NO_THROW(HINDER_EXPECTS(true, generic_error));
 }
 
-TEST_CASE("HINDER_ENSURES throws on false", "[exception]") {
+TEST(Exception, Hinder_ensuresThrowsOnFalse) {
     try {
         HINDER_ENSURES(1 == 2, generic_error);
-        FAIL("Expected exception");
+        FAIL() << "Expected exception";
     } catch (exception const& e) {
-        CHECK(e.get_as<std::string>("condition").value_or("") == "1 == 2");
-        CHECK(e.get_as<std::string>("check_type").value_or("") == "postcondition");
+        EXPECT_EQ(e.get_as<std::string>("condition").value_or(""), "1 == 2");
+        EXPECT_EQ(e.get_as<std::string>("check_type").value_or(""), "postcondition");
     }
 }
 
-TEST_CASE("HINDER_ENSURES passes on true", "[exception]") {
-    CHECK_NOTHROW(HINDER_ENSURES(true, generic_error));
+TEST(Exception, Hinder_ensuresPassesOnTrue) {
+    EXPECT_NO_THROW(HINDER_ENSURES(true, generic_error));
 }
 
-TEST_CASE("HINDER_INVARIANT throws on false", "[exception]") {
+TEST(Exception, Hinder_invariantThrowsOnFalse) {
     try {
         HINDER_INVARIANT(false, generic_error);
-        FAIL("Expected exception");
+        FAIL() << "Expected exception";
     } catch (exception const& e) {
-        CHECK(e.get_as<std::string>("condition").value_or("") == "false");
-        CHECK(e.get_as<std::string>("check_type").value_or("") == "invariant");
+        EXPECT_EQ(e.get_as<std::string>("condition").value_or(""), "false");
+        EXPECT_EQ(e.get_as<std::string>("check_type").value_or(""), "invariant");
     }
 }
 
-TEST_CASE("HINDER_INVARIANT passes on true", "[exception]") {
-    CHECK_NOTHROW(HINDER_INVARIANT(true, generic_error));
+TEST(Exception, Hinder_invariantPassesOnTrue) {
+    EXPECT_NO_THROW(HINDER_INVARIANT(true, generic_error));
 }
 
 // ============================================================================
@@ -334,115 +382,115 @@ TEST_CASE("HINDER_INVARIANT passes on true", "[exception]") {
 HINDER_DEFINE_EXCEPTION(custom_error, generic_error);
 HINDER_DEFINE_EXCEPTION(specific_error, custom_error);
 
-TEST_CASE("Custom exception types are catchable by type", "[exception]") {
-    CHECK_THROWS_AS(throw custom_error(), custom_error);
-    CHECK_THROWS_AS(throw specific_error(), specific_error);
+TEST(Exception, CustomExceptionTypesAreCatchableByType) {
+    EXPECT_THROW(throw custom_error(), custom_error);
+    EXPECT_THROW(throw specific_error(), specific_error);
 }
 
-TEST_CASE("Custom exception types are catchable by base types", "[exception]") {
-    CHECK_THROWS_AS(throw custom_error(), generic_error);
-    CHECK_THROWS_AS(throw custom_error(), exception);
-    CHECK_THROWS_AS(throw specific_error(), custom_error);
-    CHECK_THROWS_AS(throw specific_error(), generic_error);
+TEST(Exception, CustomExceptionTypesAreCatchableByBaseTypes) {
+    EXPECT_THROW(throw custom_error(), generic_error);
+    EXPECT_THROW(throw custom_error(), exception);
+    EXPECT_THROW(throw specific_error(), custom_error);
+    EXPECT_THROW(throw specific_error(), generic_error);
 }
 
-TEST_CASE("type_name() returns correct name", "[exception]") {
+TEST(Exception, Type_nameReturnsCorrectName) {
     try {
         throw custom_error();
     } catch (exception const& e) {
-        CHECK(e.type_name() == "custom_error");
+        EXPECT_EQ(e.type_name(), "custom_error");
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("what() returns type name", "[exception]") {
+TEST(Exception, WhatReturnsTypeName) {
     try {
         throw custom_error();
     } catch (std::exception const& e) {
-        CHECK(std::string(e.what()) == "custom_error");
+        EXPECT_EQ(std::string(e.what()), "custom_error");
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
 // ============================================================================
 // Edge cases
 // ============================================================================
 
-TEST_CASE("Empty exception", "[exception]") {
+TEST(Exception, EmptyException) {
     try {
         throw generic_error();
     } catch (exception const& e) {
-        CHECK(e.size() == 0);
-        CHECK(e.begin() == e.end());
+        EXPECT_EQ(e.size(), 0);
+        EXPECT_EQ(e.begin(), e.end());
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("Large string values", "[exception]") {
+TEST(Exception, LargeStringValues) {
     std::string large_string(10000, 'x');
     try {
         throw generic_error().with("large", large_string);
     } catch (exception const& e) {
         auto val = e.get_as<std::string>("large");
-        REQUIRE(val.has_value());
-        CHECK(val->size() == 10000);
+        ASSERT_TRUE(val.has_value());
+        EXPECT_EQ(val->size(), 10000);
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("Unicode strings", "[exception]") {
+TEST(Exception, UnicodeStrings) {
     try {
         throw generic_error()
             .with("emoji", "Hello 🌍")
             .with("chinese", "你好世界")
             .with("mixed", "Héllo Wörld");
     } catch (exception const& e) {
-        CHECK(e.get_as<std::string>("emoji").value_or("") == "Hello 🌍");
-        CHECK(e.get_as<std::string>("chinese").value_or("") == "你好世界");
-        CHECK(e.get_as<std::string>("mixed").value_or("") == "Héllo Wörld");
+        EXPECT_EQ(e.get_as<std::string>("emoji").value_or(""), "Hello 🌍");
+        EXPECT_EQ(e.get_as<std::string>("chinese").value_or(""), "你好世界");
+        EXPECT_EQ(e.get_as<std::string>("mixed").value_or(""), "Héllo Wörld");
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
-TEST_CASE("message() with format arguments", "[exception]") {
+TEST(Exception, MessageWithFormatArguments) {
     try {
         throw generic_error().message("Value {} exceeds limit {}", 42, 100);
     } catch (exception const& e) {
-        CHECK(e.get_as<std::string>("message").value_or("") == "Value 42 exceeds limit 100");
+        EXPECT_EQ(e.get_as<std::string>("message").value_or(""), "Value 42 exceeds limit 100");
         return;
     }
-    FAIL("Expected exception to be thrown");
+    FAIL() << "Expected exception to be thrown";
 }
 
 // ============================================================================
 // value_to_string
 // ============================================================================
 
-TEST_CASE("value_to_string monostate", "[exception]") {
-    CHECK(value_to_string(std::monostate{}) == "");
+TEST(Exception, Value_to_stringMonostate) {
+    EXPECT_EQ(value_to_string(std::monostate{}), "");
 }
 
-TEST_CASE("value_to_string bool", "[exception]") {
-    CHECK(value_to_string(true) == "true");
-    CHECK(value_to_string(false) == "false");
+TEST(Exception, Value_to_stringBool) {
+    EXPECT_EQ(value_to_string(true), "true");
+    EXPECT_EQ(value_to_string(false), "false");
 }
 
-TEST_CASE("value_to_string integers", "[exception]") {
-    CHECK(value_to_string(std::int64_t{-42}) == "-42");
-    CHECK(value_to_string(std::uint64_t{42}) == "42");
+TEST(Exception, Value_to_stringIntegers) {
+    EXPECT_EQ(value_to_string(std::int64_t{-42}), "-42");
+    EXPECT_EQ(value_to_string(std::uint64_t{42}), "42");
 }
 
-TEST_CASE("value_to_string double", "[exception]") {
-    CHECK(value_to_string(3.14) == "3.14");
+TEST(Exception, Value_to_stringDouble) {
+    EXPECT_EQ(value_to_string(3.14), "3.14");
 }
 
-TEST_CASE("value_to_string string", "[exception]") {
-    CHECK(value_to_string(std::string{"hello"}) == "hello");
+TEST(Exception, Value_to_stringString) {
+    EXPECT_EQ(value_to_string(std::string{"hello"}), "hello");
 }
 
 // ============================================================================
@@ -451,36 +499,36 @@ TEST_CASE("value_to_string string", "[exception]") {
 
 #ifndef NDEBUG
 
-TEST_CASE("HINDER_ASSERT throws on false condition", "[exception][assert]") {
-    CHECK_THROWS_AS(HINDER_ASSERT(1 == 0, "condition is false"), assertion_error);
+TEST(Assert, HinderassertThrowsOnFalseCondition) {
+    EXPECT_THROW(HINDER_ASSERT(1 == 0, "condition is false"), assertion_error);
 }
 
-TEST_CASE("HINDER_ASSERT passes on true condition", "[exception][assert]") {
-    CHECK_NOTHROW(HINDER_ASSERT(1 == 1, "condition is true"));
+TEST(Assert, HinderassertPassesOnTrueCondition) {
+    EXPECT_NO_THROW(HINDER_ASSERT(1 == 1, "condition is true"));
 }
 
-TEST_CASE("HINDER_ASSERT captures condition and message", "[exception][assert]") {
+TEST(Assert, HinderassertCapturesConditionAndMessage) {
     try {
         int x = 42;
         HINDER_ASSERT(x == 0, "expected {}, got {}", 0, x);
-        FAIL("Expected assertion_error");
+        FAIL() << "Expected assertion_error";
     } catch (assertion_error const& e) {
-        CHECK(e.get_as<std::string>("condition").value_or("") == "x == 0");
-        CHECK(e.get_as<std::string>("check_type").value_or("") == "assertion");
+        EXPECT_EQ(e.get_as<std::string>("condition").value_or(""), "x == 0");
+        EXPECT_EQ(e.get_as<std::string>("check_type").value_or(""), "assertion");
         auto msg = e.get_as<std::string>("message").value_or("");
-        CHECK_THAT(msg, Contains("expected 0"));
-        CHECK_THAT(msg, Contains("got 42"));
+        EXPECT_THAT(msg, HasSubstr("expected 0"));
+        EXPECT_THAT(msg, HasSubstr("got 42"));
     }
 }
 
-TEST_CASE("HINDER_ASSERT captures source location", "[exception][assert]") {
+TEST(Assert, HinderassertCapturesSourceLocation) {
     try {
         HINDER_ASSERT(false, "test assertion");
-        FAIL("Expected assertion_error");
+        FAIL() << "Expected assertion_error";
     } catch (assertion_error const& e) {
         auto const& loc = e.location();
-        CHECK(std::string(loc.file_name()).find("exception_tests.cpp") != std::string::npos);
-        CHECK(loc.line() > 0);
+        EXPECT_NE(std::string(loc.file_name()).find("exception_tests.cpp"), std::string::npos);
+        EXPECT_TRUE(loc.line() > 0);
     }
 }
 
